TimerReceiver: Correctly show the remaining time for 100 hours
If you open the Clock countdown time and input 99h99m99s it will
actually translate to 100 hours, 40 minutes and 39 seconds as
expected.

When leaving the app it will show a notification with the remaining
time and if there are more than 100 hours it will fail to show
that part and only show "40 minutes remaining".

This patch fixes allows for 100 as a valid input.

Change-Id: Ib3621691aaa3a449262cb9e3601d39283bb808f8